Move make_all_safe to Board
