+++ j5/base_robot.py
@@ -1,7 +1,5 @@
 """A base class for robots."""
 
-import atexit
-
 from j5.boards import Board
 
 
@@ -10,11 +8,4 @@ class BaseRobot:
 
     def make_safe(self):
         """Make this robot safe."""
-        BaseRobot.make_all_safe()
-
-    @staticmethod
-    @atexit.register
-    def make_all_safe():
-        """Make all boards safe."""
-        for board in Board.BOARDS:
-            board.make_safe()
+        Board.make_all_safe()
